refactor: remove top-level `pkg` directory
